requirements.txt
loguru
mipa==0.4.1
mipac==0.6.2
python-dotenv
